Added functions for pytests to test the classical particle radius for ions, protons and for proper printing of values.
